Track bonded terms for protocell
Testing with Noah to determine if some of the bonded terms fail to be
evaluated as expected. Define macro DEBUG_PROTOCELL to track the bonded
terms whose atom indices are all within the given range defined in
common.h. The compute objects will range check the indices and print out
the bonded terms, so best to just run for 0 steps.

To use this patch, add to Make.config:
  CXXOPTS += -DDEBUG_PROTOCELL

The index range being checked was specified by Noah:
  PRMIN  23749759
  PRMAX  23749892
The default can be changed when building, e.g.,
  CXXOPTS += -D PRMIN=23749760

Change-Id: If747b120d54a8ec12dd8e15a26886ad6e0609347