Streamlined integration for SOA data
Implementing the subroutines needed for the Sequencer integrator using
the SOA data structures.  The code is presently inactive and only about
halfway finished.  The finished version will test whether only the
reduced set of supported options are being used for integration and then
set a flag (probably in SimParameters) that is tested to choose between
calling Sequencer::integrate_SOA() or falling back to
Sequencer::integrate().

Change-Id: I8b53113695f4d49a8c84d2b9544d024ebe1beb95