Initial version of streamlined integrator
Initial version of integrate_SOA() in Sequencer is activated.
Use it by setting "SOAintegrate on" in config file.  Testing has
demonstrated correct results on alanin, ApoA1, and STMV.

Performance is not yet improved due to the absence of any deep
modifications to the rigid bond constraints or the random number
generation.  For now, the rattle1_SOA() copies back to AOS before
calling the unmodified HomePatch routines.

Change-Id: If080f2720370e46f9e6d4873f310eead6e31fd0b