Update SOA rigid bond constraints
HomePatch::rattle1_SOA() is revised to avoid extra buffering and also
to eliminate calculations on fixed atoms, which are not supported by
this code path.  Solvent is now placed after solute to allow calling a
revised SETTLE routine that performs a single loop over all water
molecules.  The routine to initialize the rattle lists in
HomePatch::buildRattleList_SOA() is now factored out of the main rattle
routine, performed earlier at HomePatch setup or reinitialization or
right after atom migration, to avoid extra testing during rattle1_SOA().

The "isWater" flag for atoms, defined in CompAtom, is now being
initialized in the WorkDistrib.C routines for both the standard and
memory optimized versions.  It specifically designates for a water
molecule that the SETTLE conditions hold.  The field is used for the
separation of solvent from solute.

The SOA code path is getting correct results, but performance is still
not as good as for the AOS code path.  Separation of solvent is done
into two additional FullAtom arrays, that are then copied back into the
main FullAtom array before copying to SOA form.  Performance should be
improved by avoiding the explicit formation of a sorted FullAtom array
until either atom migration or collection.

Other small updates.  The code enabled by the NAMD_SeparateWaters macro
is broken, causing an infinite loop, as explained by a comment.  Another
Doxygen comment is placed at the start of the Molecule class definition.
Additional checks for SOAintegratOn now test against pressureProfileOn
and to make sure that WAT_TIP3 is set.

Still need to remove dead and debugging codes.

Change-Id: I6f8820ba0ce7d5a7ea233113f95608a61aad0b17