Revert "Double wide FEP sampling"
This reverts commit b0b6ab892c4f8f14fe6afa91605ccda635891cc2.

Brian Radak requested that this earlier commit be reverted.
The implementation does not work well with recent changes in
master to WCA and PME and is causing a conflict when trying to
synchronize with the most recent patch that fixes LJcorrection
for alchemical coupling. Brian asked for a clean slate to rewrite
the double-wide sampling codes. Also, the ParseFEP tool is not
yet ready to take advantage of the extra output generated.

Change-Id: Ib3ee8808bc6cb605f1e7294bcfb3869fbe5366db