Revise SOA data structure and clean up code
Memory for SOA data structure is now allocated as one large storage
block containing consecutive arrays.  The array lengths are padded to
the next multiple of 32 for vectorization compatible with CUDA.  What
had been a ResizeArray for each data element is now a pointer into the
storage block.  Storage can be resized larger if the number of atoms in
the patch increases, where existing data is preserved, like realloc.
Use of a single storage block facilitates copying data to the GPU.

The big storage block is still using ResizeArray for the convenience
of easily obtaining a memory aligned allocation, but this will have
to change if we want PatchDataSOA to also be able to reside on the GPU.

SOA code path has been cleaned up of dead code to help make it easier
to read.  Comments have been added, particularly for PatchDataSOA
structure, along with many XXX comments indicating things to consider.

Change-Id: I9748eac0d28eafd3780ba317c97264582351e38c